pytorch/requirements.txt
matplotlib
nltk
numpy
Pillow
argparse
torch
torchvision
pycocotools
